Refactor CI change detection gate
Split workflow validation, provenance handling, project graph checks, shell execution, and scenario assertions into explicit owners while preserving the existing file-based command.


Copilot-Session: cd7e27ec-810c-462f-92be-3e539961015b
