+++ docs/decompiler-correctness-pipeline.md
@@ -302,13 +302,14 @@ notification was ever sent. Detection latency was unbounded, not weekly
 The `decompiler-gates` CI job closes that hole. Source, test, and tool projects
 run it by default, except for measured false positives in
 `eng/decompiler-gate-skip-projects.txt`. The initial exemptions are the CLI and
-its tests, which do not feed the gate. `test-ci-change-detection.cs` asserts
-that no exempted project appears in MSBuild's evaluated project-reference
-closure rooted at `ILInspector.Decompiler.Tests`. New projects therefore run
-the gate without a list update; an unreadable or invalid skip list exempts
-nothing. Global build inputs and the gate's own scripts and pins remain
-explicit triggers. The job runs separately so it never serializes with the hot
-`test` lane, and executes `--gate pre-merge`.
+its tests, which do not feed the gate. `DecompilerProjectGraphPolicy` in the
+`eng/CiChangeDetection` gate asserts that no exempted project appears in
+MSBuild's evaluated project-reference closure rooted at
+`ILInspector.Decompiler.Tests`. New projects therefore run the gate without a
+list update; an unreadable or invalid skip list exempts nothing. Global build
+inputs and the gate's own scripts and pins remain explicit triggers. The job
+runs separately so it never serializes with the hot `test` lane, and executes
+`--gate pre-merge`.
 
 ```bash
 dotnet run --project src/ILInspector.Decompiler.Tests -c Release -- \
