+++ eng/Directory.Build.props
@@ -3,14 +3,14 @@
   <Import Project="$([MSBuild]::GetPathOfFileAbove('Directory.Build.props', '$(MSBuildThisFileDirectory)../'))" />
 
   <PropertyGroup>
-    <!-- eng/ holds build-time file-based apps: sweep preparation, gate checks,
-         and measurement. They are outside the CLI's reference closure for the
-         same reason the harnesses under tools/ are (#3319) — each is its own
-         entry point reporting to the operator who ran it, cannot reach
+    <!-- eng/ holds build-time file-based apps and their supporting gate code:
+         sweep preparation, gate checks, and measurement. They are outside the
+         CLI's reference closure for the same reason the harnesses under tools/
+         are (#3319) — each reports to the operator who ran it, cannot reach
          dotnet-inspect's CommandError, and would become a consumer of the
          product surface it exists to check if it were routed through one.
 
-         Declared here rather than in each file because a file-based app has no
+         Declared here rather than in each app because a file-based app has no
          project file to carry the property. The extension guard in
          Directory.Build.targets does not cover these: a file-based app builds
          through a generated project whose MSBuildProjectExtension is .csproj,
